sw_services: xilsecure: Fix MISRA C Violation of Rule 10.3
MISRA C violation 10.3 : The value of an expression shall not be assigned to an object
with a narrower or of a different essential type

Soln : Fixed by proper type casting of the operands.

Acked-by: Harsha <harsha.harsha@xilinx.com>

+++ lib/sw_services/xilsecure/src/common/versal_common/server/xsecure_aes_ipihandler.c
@@ -568,7 +568,7 @@ static int XSecure_AesPerformOperation(u32 SrcAddrLow, u32 SrcAddrHigh)
 	Status = XST_FAILURE;
 	if (AesParams.OperationId == (u32)XSECURE_ENCRYPT) {
 		if (XPlmi_IsKatRan(XPLMI_SECURE_AES_ENC_KAT_MASK) != TRUE) {
-			Status = XSECURE_ERR_KAT_NOT_EXECUTED;
+			Status = (int)XSECURE_ERR_KAT_NOT_EXECUTED;
 			goto END;
 		}
 		/**<AES Encrypt Init*/
